Constructor for StoreMap

+++ src/main/java/sa/Customer.java
@@ -8,7 +8,7 @@ public class Customer {
 	public boolean[] products;  //Products this costumer wants
 	// (ex: [true, false, false, true] = costumer wants products 0 and 3)
 	
-	public Customer(int posx, int posy, boolean[] products) {
+	public Customer(float posx, float posy, boolean[] products) {
 		this.posx = posx;
 		this.posy = posy;
 		this.products = products;
